project(factories): move serviceFactory to new folder
